# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger API documentation for test task",
        "contact": {
            "name": "Dmytro",
            "email": "dm.kopylets@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/articles/{article}/comments": {
            "get": {
                "tags": [
                    "comments"
                ],
                "summary": "List of comments to the article",
                "operationId": "getCommentsList",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Articles id to show list",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "comments not found"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/articles/{article_id}/comments/add-first": {
            "post": {
                "tags": [
                    "comments"
                ],
                "summary": "add firstLevel comment to the article",
                "operationId": "addFirstComment",
                "parameters": [
                    {
                        "name": "article_id",
                        "in": "path",
                        "description": "Articles id to comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Users id to comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "body",
                        "in": "query",
                        "description": "Text comment",
                        "schema": {
                            "type": "text"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "comments not found"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/articles/{article_id}/comments/add-reply": {
            "post": {
                "tags": [
                    "comments"
                ],
                "summary": "add comment to comment",
                "operationId": "addReply",
                "parameters": [
                    {
                        "name": "article_id",
                        "in": "path",
                        "description": "Articles id to comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Users id to comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "parent_id",
                        "in": "path",
                        "description": "id parent's comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "body",
                        "in": "query",
                        "description": "Text comment",
                        "schema": {
                            "type": "text"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "comments not found"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/articles/comments/store": {
            "post": {
                "tags": [
                    "comments"
                ],
                "summary": "main action",
                "operationId": "saveComment",
                "requestBody": {
                    "description": "Client side request to store comment",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "user_id": {
                                                "type": "integer"
                                            },
                                            "article_id": {
                                                "type": "integer"
                                            },
                                            "body": {
                                                "type": "string"
                                            },
                                            "parent_id": {
                                                "type": "integer"
                                            },
                                            "position": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/articles/{article}/comments/reSortByDate": {
            "get": {
                "tags": [
                    "comments"
                ],
                "summary": "Sort the list of article comments by the date of creation",
                "operationId": "getCommentsListSortedByDate",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Articles id to show list",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Direction for sorting the list",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ }"
                    }
                }
            }
        },
        "/api/v1/articles/{article}/comments/reSortByName": {
            "get": {
                "tags": [
                    "comments"
                ],
                "summary": "Sort the list of article comments by user name",
                "operationId": "getCommentsListSortedByName",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Articles id to show list",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Direction for sorting the list",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ }"
                    }
                }
            }
        },
        "/api/v1/articles/{article}/comments/reSortByEmail": {
            "get": {
                "tags": [
                    "comments"
                ],
                "summary": "Sort the list of article comments by user email",
                "operationId": "getCommentsListSortedByEmail",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Articles id to show list",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Direction for sorting the list",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ }"
                    }
                }
            }
        },
        "/api/v1/articles": {
            "get": {
                "tags": [
                    "articles"
                ],
                "summary": "Articles listing",
                "operationId": "getArticlesList",
                "parameters": [
                    {
                        "name": "wantedAuthor",
                        "in": "query",
                        "description": "part of the user name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "wantedTitle",
                        "in": "query",
                        "description": "part of the article's title",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "articles not found"
                    }
                }
            }
        }
    }
}